Rewrite string to formula conversion
Redesigned the algorithm that converts a string into a formula,
and therefore removed all "private" methods, and added new ones.

Modified the <LinearFormula.__init__> method so that it can be
initialized with a string, a dict, an int, or two lists of multipliers
and variables.

Organised the methods into sections and separated the sections with
'#-------------------...' lines

Added tests that test the new capabilities of the <LinearFormula> class.